Create a POC of workflow graph for rollout sequencing
This is minimal front-end POC that we plan to show during UX research. I've tried still to make code more or less clean, but we will do a proper design for this UI and will do a full refactor of this code once we'll start the implementation. Thus, there are some pieces missing compared to usual code delivery (e.g. extensive unit testing coverage).

PiperOrigin-RevId: 598574206