#dag Support external translation messages
Context: The DAG currently supports setting i18n messages in templates but had a fake function for actual translations in component code. This change enables setting translations from the host code. It is currently only used in the node-state-badge (used in Vertex only) component, so it's not introduced in the main component or the dag-wrapper.

In the future, if we need other translation messages we can add a similar `translations` input where needed.

The translations are provided internally with goog.getMsg() and can be provided arbitrarily by OSS users for each translation key based on the current locale.

Other options considered:
1. Provide a dictionary only and use it directly: While a bit simpler, this does not allow accessing translation values across the codebase (hence the service).
2. Use a Provider instead of an @Input(): The component can be used as a web-component where Angular Providers can not be customized.
3. Use a template-based @switch: Could be a bit simpler for now, but we actually need a way to translate strings in the component code for future-proofing the DAG.

PiperOrigin-RevId: 604929824