Currently, Angular has a bug in which both `foo` and `this.foo` in a template interpolation are interpreted as reading a template context variable, if one is defined.
The Angular team intends to [fix this bug](https://github.com/angular/angular/pull/55183), after which `foo` will still refer to the template variable, but `this.foo` will always properly refer to the class.

Your code attempts to read a template variable with `this.`. By removing `this.` you will be forward compatible with our fix.

PiperOrigin-RevId: 643899468