LSC — Explicitly mark AOT incompatible targets via `jit: true (shard 2)
This LSC will add `jit: true` to all `@Directive`, `@Component` and `@NgModule` classes that are not AOT compatible.

Commonly in Google3, TypeScript test files are part of a `ts_library` or a `ng_module w/ testonly=True`. These targets simply run the TypeScript compiler from Microsoft, without any specific Angular processing. This allows code to exist that isn't necessarily compatible with Angular's compiler that strictly checks e.g. components

This LSC adds `jit: true` to those components to clearly indicate that those components are not processed by the Angular compiler. This is a no-op because the Angular compiler doesn't even run on these files (as said above). In the future, we would like to change this, and need clear indication of what is incompatible or not. This is what the `jit: true` flag does.  The `jit: true` flag basically tells the Angular compiler to ignore these components, so that their behavior is unchanged and everything continues to work like before.

For additional details, see: https://docs.google.com/document/u/0/d/1ox4atCJldWWDXlaYgwM-hU8BNsTpKNW7gx8OfZ0HtRY/edit

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:655462973:BASE:655450343:1721825823310:897c078a
PiperOrigin-RevId: 655573581