Introduce labels to groups in the workflow graph. This is useful for distinguishing between different groups when control nodes are not present.
The labels are displayed in a small box at the top of the group. The box is positioned so that it does not overlap with the control node or the expand/collapse button.

The labels are optional. If a group does not have a label, then no label will be displayed.

PiperOrigin-RevId: 698365593