Fetch execution count per node for async run
### Why
This will give users more accurate view of the graph of an async run by exposing the actual number of executions per node for async run without increasing the overall latency to render the page. Note that displaying the number of artifacts is much more expensive as it requires us to fetch events, which takes time that is not much different from fetching the whole data.

### Testing
* Screencast: https://screencast.googleplex.com/cast/NDk2Mjc1NDE0MTA5Mzg4OHw2MWRlOGVlYS0zZA

PiperOrigin-RevId: 708111788