Fix workflow graph nodes are not centered within a group
Context: We calculate the graph height by default by calculating the topmost/bottommost node's Y coordinate (center point) and then adding half of the default node height to it (to adjust for the center point) and then adding the default paddings and margins. The error happens when the nodes are custom -> their height does not equal the default node height. To fix this, this CL modifies the topmost/bottommost point getter function to also calculate with an offset.
PiperOrigin-RevId: 714032871