Move `cmIcon` from `NodeIcon` to `Node.options`.
This change refactors how the `SVGSVGElement` for custom icons is stored in the workflow graph. Instead of being part of the `NodeIcon` interface, the `cmIcon` is now stored in the `options` field of the `Node` object. This provides better separation, as `NodeIcon` focuses on icon metadata, while `options` can hold additional node-specific data like the `cmIcon`.

PiperOrigin-RevId: 820772263